gasOptimizations for BathPair:

+++ contracts/peripheral_contracts/USDCWithFaucet.sol
@@ -86,7 +86,8 @@ contract USDCWithFaucet is LibNote, ERC20 {
     // bytes32 public constant PERMIT_TYPEHASH = keccak256("Permit(address holder,address spender,uint256 nonce,uint256 expiry,bool allowed)");
     bytes32 public constant PERMIT_TYPEHASH =
         0xea2aa0a1be11a07ed86d755c93467f4f82362b452371d1ba94d1715123511acb;
-    uint public timeDelay;
+    uint256 public timeDelay;
+
     constructor(
         uint256 chainId_,
         address _admin,
@@ -124,10 +125,12 @@ contract USDCWithFaucet is LibNote, ERC20 {
         require(admin == msg.sender);
         _mint(msg.sender, 10000e18);
     }
-    function setTimeDelay(uint _timeDelay) external {
+
+    function setTimeDelay(uint256 _timeDelay) external {
         require(admin == msg.sender);
         timeDelay = _timeDelay;
     }
+
     // --- Alias ---
     function push(address usr, uint256 wad) external {
         transferFrom(msg.sender, usr, wad);
